+++ CHANGELOG.md
@@ -11,6 +11,9 @@
     - file.get
     - file.get_stream
     - read.json
+  - snowflake.snowpark.functions
+    - udf
+- Added the function `DataFrame.write.csv` to unload data from a ``DataFrame`` into one or more CSV files in a stage.
 
 ## 1.14.0 (2024-03-20)